+++ src/include/dof.h
@@ -3,12 +3,12 @@
 class Dof
 {
 private:
-    float value;
+    double value;
     static unsigned int global_dof_id_counter;
 public:
     const unsigned int id;
     unsigned int get_global_dof_id_counter();
-    float get_value();
+    double get_value();
     Dof();
     ~Dof();
 };
