+++ src/src/C3D10.cpp
@@ -4,19 +4,19 @@
 
 void C3D10::calculate_Ke(){
     setup_coord();
-    Eigen::Matrix<float,3,3> J;
-    Eigen::Matrix<float,6,30> B;
+    Eigen::Matrix<double,3,3> J;
+    Eigen::Matrix<double,6,30> B;
     
     // size(dNdXhi) = ngp*ngp, 4 vectors of (1x4)
-    Eigen::Matrix<float,3,10> dNdXhidEtadMy;
-    Eigen::Matrix<float,3,10> dNdxdydz;
+    Eigen::Matrix<double,3,10> dNdXhidEtadMy;
+    Eigen::Matrix<double,3,10> dNdxdydz;
 
     // init Ke zero
     Ke.setZero();
-    float dN1dXhi, dN2dXhi, dN3dXhi, dN4dXhi, dN5dXhi, dN6dXhi, dN7dXhi, dN8dXhi, dN9dXhi, dN10dXhi;
-    float dN1dEta, dN2dEta, dN3dEta, dN4dEta, dN5dEta, dN6dEta, dN7dEta, dN8dEta, dN9dEta, dN10dEta;
-    float dN1dMy, dN2dMy, dN3dMy, dN4dMy, dN5dMy, dN6dMy, dN7dMy, dN8dMy, dN9dMy, dN10dMy;
-    float xhi,eta,my,w;
+    double dN1dXhi, dN2dXhi, dN3dXhi, dN4dXhi, dN5dXhi, dN6dXhi, dN7dXhi, dN8dXhi, dN9dXhi, dN10dXhi;
+    double dN1dEta, dN2dEta, dN3dEta, dN4dEta, dN5dEta, dN6dEta, dN7dEta, dN8dEta, dN9dEta, dN10dEta;
+    double dN1dMy, dN2dMy, dN3dMy, dN4dMy, dN5dMy, dN6dMy, dN7dMy, dN8dMy, dN9dMy, dN10dMy;
+    double xhi,eta,my,w;
     auto D = pid->get_mid()->D_3D_linear_continuum_mechanics;
     for (unsigned int i = 0; i < gauss_points->size(); i++)
     {
@@ -67,12 +67,12 @@ void C3D10::calculate_Ke(){
         // find shape functions derivate matrix wrt x, y & z
         dNdxdydz = J.inverse()*dNdXhidEtadMy;
         // construct B matrix
-        B << dNdxdydz(0,0),0.0f,0.0f,dNdxdydz(0,1),0.0f,0.0f,dNdxdydz(0,2),0.0f,0.0f,dNdxdydz(0,3),0.0f,0.0f,dNdxdydz(0,4),0.0f,0.0f,dNdxdydz(0,5),0.0f,0.0f,dNdxdydz(0,6),0.0f,0.0f,dNdxdydz(0,7),0.0f,0.0f,dNdxdydz(0,8),0.0f,0.0f,dNdxdydz(0,9),0.0f,0.0f,
-             0.0f,dNdxdydz(1,0),0.0f,0.0f,dNdxdydz(1,1),0.0f,0.0f,dNdxdydz(1,2),0.0f,0.0f,dNdxdydz(1,3),0.0f,0.0f,dNdxdydz(1,4),0.0f,0.0f,dNdxdydz(1,5),0.0f,0.0f,dNdxdydz(1,6),0.0f,0.0f,dNdxdydz(1,7),0.0f,0.0f,dNdxdydz(1,8),0.0f,0.0f,dNdxdydz(1,9),0.0f,
-             0.0f,0.0f,dNdxdydz(2,0),0.0f,0.0f,dNdxdydz(2,1),0.0f,0.0f,dNdxdydz(2,2),0.0f,0.0f,dNdxdydz(2,3),0.0f,0.0f,dNdxdydz(2,4),0.0f,0.0f,dNdxdydz(2,5),0.0f,0.0f,dNdxdydz(2,6),0.0f,0.0f,dNdxdydz(2,7),0.0f,0.0f,dNdxdydz(2,8),0.0f,0.0f,dNdxdydz(2,9),
-             dNdxdydz(1,0),dNdxdydz(0,0),0.0f,dNdxdydz(1,1),dNdxdydz(0,1),0.0f,dNdxdydz(1,2),dNdxdydz(0,2),0.0f,dNdxdydz(1,3),dNdxdydz(0,3),0.0f,dNdxdydz(1,4),dNdxdydz(0,4),0.0f,dNdxdydz(1,5),dNdxdydz(0,5),0.0f,dNdxdydz(1,6),dNdxdydz(0,6),0.0f,dNdxdydz(1,7),dNdxdydz(0,7),0.0f,dNdxdydz(1,8),dNdxdydz(0,8),0.0f,dNdxdydz(1,9),dNdxdydz(0,9),0.0f,
-             0.0f,dNdxdydz(2,0),dNdxdydz(1,0),0.0f,dNdxdydz(2,1),dNdxdydz(1,1),0.0f,dNdxdydz(2,2),dNdxdydz(1,2),0.0f,dNdxdydz(2,3),dNdxdydz(1,3),0.0f,dNdxdydz(2,4),dNdxdydz(1,4),0.0f,dNdxdydz(2,5),dNdxdydz(1,5),0.0f,dNdxdydz(2,6),dNdxdydz(1,6),0.0f,dNdxdydz(2,7),dNdxdydz(1,7),0.0f,dNdxdydz(2,8),dNdxdydz(1,8),0.0f,dNdxdydz(2,9),dNdxdydz(1,9),
-             dNdxdydz(2,0),0.0f,dNdxdydz(0,0),dNdxdydz(2,1),0.0f,dNdxdydz(0,1),dNdxdydz(2,2),0.0f,dNdxdydz(0,2),dNdxdydz(2,3),0.0f,dNdxdydz(0,3),dNdxdydz(2,4),0.0f,dNdxdydz(0,4),dNdxdydz(2,5),0.0f,dNdxdydz(0,5),dNdxdydz(2,6),0.0f,dNdxdydz(0,6),dNdxdydz(2,7),0.0f,dNdxdydz(0,7),dNdxdydz(2,8),0.0f,dNdxdydz(0,8),dNdxdydz(2,9),0.0f,dNdxdydz(0,9);
+        B << dNdxdydz(0,0),0.d,0.d,dNdxdydz(0,1),0.d,0.d,dNdxdydz(0,2),0.d,0.d,dNdxdydz(0,3),0.d,0.d,dNdxdydz(0,4),0.d,0.d,dNdxdydz(0,5),0.d,0.d,dNdxdydz(0,6),0.d,0.d,dNdxdydz(0,7),0.d,0.d,dNdxdydz(0,8),0.d,0.d,dNdxdydz(0,9),0.d,0.d,
+             0.d,dNdxdydz(1,0),0.d,0.d,dNdxdydz(1,1),0.d,0.d,dNdxdydz(1,2),0.d,0.d,dNdxdydz(1,3),0.d,0.d,dNdxdydz(1,4),0.d,0.d,dNdxdydz(1,5),0.d,0.d,dNdxdydz(1,6),0.d,0.d,dNdxdydz(1,7),0.d,0.d,dNdxdydz(1,8),0.d,0.d,dNdxdydz(1,9),0.d,
+             0.d,0.d,dNdxdydz(2,0),0.d,0.d,dNdxdydz(2,1),0.d,0.d,dNdxdydz(2,2),0.d,0.d,dNdxdydz(2,3),0.d,0.d,dNdxdydz(2,4),0.d,0.d,dNdxdydz(2,5),0.d,0.d,dNdxdydz(2,6),0.d,0.d,dNdxdydz(2,7),0.d,0.d,dNdxdydz(2,8),0.d,0.d,dNdxdydz(2,9),
+             dNdxdydz(1,0),dNdxdydz(0,0),0.d,dNdxdydz(1,1),dNdxdydz(0,1),0.d,dNdxdydz(1,2),dNdxdydz(0,2),0.d,dNdxdydz(1,3),dNdxdydz(0,3),0.d,dNdxdydz(1,4),dNdxdydz(0,4),0.d,dNdxdydz(1,5),dNdxdydz(0,5),0.d,dNdxdydz(1,6),dNdxdydz(0,6),0.d,dNdxdydz(1,7),dNdxdydz(0,7),0.d,dNdxdydz(1,8),dNdxdydz(0,8),0.d,dNdxdydz(1,9),dNdxdydz(0,9),0.d,
+             0.d,dNdxdydz(2,0),dNdxdydz(1,0),0.d,dNdxdydz(2,1),dNdxdydz(1,1),0.d,dNdxdydz(2,2),dNdxdydz(1,2),0.d,dNdxdydz(2,3),dNdxdydz(1,3),0.d,dNdxdydz(2,4),dNdxdydz(1,4),0.d,dNdxdydz(2,5),dNdxdydz(1,5),0.d,dNdxdydz(2,6),dNdxdydz(1,6),0.d,dNdxdydz(2,7),dNdxdydz(1,7),0.d,dNdxdydz(2,8),dNdxdydz(1,8),0.d,dNdxdydz(2,9),dNdxdydz(1,9),
+             dNdxdydz(2,0),0.d,dNdxdydz(0,0),dNdxdydz(2,1),0.d,dNdxdydz(0,1),dNdxdydz(2,2),0.d,dNdxdydz(0,2),dNdxdydz(2,3),0.d,dNdxdydz(0,3),dNdxdydz(2,4),0.d,dNdxdydz(0,4),dNdxdydz(2,5),0.d,dNdxdydz(0,5),dNdxdydz(2,6),0.d,dNdxdydz(0,6),dNdxdydz(2,7),0.d,dNdxdydz(0,7),dNdxdydz(2,8),0.d,dNdxdydz(0,8),dNdxdydz(2,9),0.d,dNdxdydz(0,9);
         // compute Gauss point contribution to element Ke matrix
         Ke = Ke + w*(B.transpose()*D*B*detJ.back());
     }
@@ -80,11 +80,11 @@ void C3D10::calculate_Ke(){
 }
 void C3D10::calculate_Me(){
     setup_coord();
-    Eigen::Matrix<float,3,30> N;
+    Eigen::Matrix<double,3,30> N;
     // init Ke zero
     Me.setZero();
-    float N1,N2,N3,N4,N5,N6,N7,N8, N9, N10;
-    float xhi,eta,my,w;
+    double N1,N2,N3,N4,N5,N6,N7,N8, N9, N10;
+    double xhi,eta,my,w;
     for (unsigned int i = 0; i < gauss_points->size(); i++)
     {
         xhi = gauss_points->at(i).at(0);
@@ -92,19 +92,19 @@ void C3D10::calculate_Me(){
         my  = gauss_points->at(i).at(2);
         w   = gauss_weights->at(i);
         // shape functions
-        N1 = eta*(2*eta - 1.0f);
-        N2 = my*(2*my - 1.0f);
-        N3 = (1.0f - xhi - eta - my)*(1 - 2*xhi - 2*eta - 2*my);
-        N4 = xhi*(2*xhi - 1.0f);
+        N1 = eta*(2*eta - 1.d);
+        N2 = my*(2*my - 1.d);
+        N3 = (1.d - xhi - eta - my)*(1 - 2*xhi - 2*eta - 2*my);
+        N4 = xhi*(2*xhi - 1.d);
         N5 = 4*eta*my;
-        N6 = 4*my*(1.0f - xhi - eta - my);
-        N7 = 4*eta*(1.0f - xhi - eta - my);
+        N6 = 4*my*(1.d - xhi - eta - my);
+        N7 = 4*eta*(1.d - xhi - eta - my);
         N8 = 4*xhi*eta;
         N9 = 4*xhi*my;
-        N10= 4*xhi*(1.0f - xhi - eta - my);
-        N << N1,0.0f,0.0f,N2,0.0f,0.0f,N3,0.0f,0.0f,N4,0.0f,0.0f,N5,0.0f,0.0f,N6,0.0f,0.0f,N7,0.0f,0.0f,N8,0.0f,0.0f,N9,0.0f,0.0f,N10,0.0f,0.0f,
-             0.0f,N1,0.0f,0.0f,N2,0.0f,0.0f,N3,0.0f,0.0f,N4,0.0f,0.0f,N5,0.0f,0.0f,N6,0.0f,0.0f,N7,0.0f,0.0f,N8,0.0f,0.0f,N9,0.0f,0.0f,N10,0.0f,
-             0.0f,0.0f,N1,0.0f,0.0f,N2,0.0f,0.0f,N3,0.0f,0.0f,N4,0.0f,0.0f,N5,0.0f,0.0f,N6,0.0f,0.0f,N7,0.0f,0.0f,N8,0.0f,0.0f,N9,0.0f,0.0f,N10;
+        N10= 4*xhi*(1.d - xhi - eta - my);
+        N << N1,0.d,0.d,N2,0.d,0.d,N3,0.d,0.d,N4,0.d,0.d,N5,0.d,0.d,N6,0.d,0.d,N7,0.d,0.d,N8,0.d,0.d,N9,0.d,0.d,N10,0.d,0.d,
+             0.d,N1,0.d,0.d,N2,0.d,0.d,N3,0.d,0.d,N4,0.d,0.d,N5,0.d,0.d,N6,0.d,0.d,N7,0.d,0.d,N8,0.d,0.d,N9,0.d,0.d,N10,0.d,
+             0.d,0.d,N1,0.d,0.d,N2,0.d,0.d,N3,0.d,0.d,N4,0.d,0.d,N5,0.d,0.d,N6,0.d,0.d,N7,0.d,0.d,N8,0.d,0.d,N9,0.d,0.d,N10;
         Me = Me + w*(pid->get_mid()->get_density()*N.transpose()*N*detJ.at(i));
     }
