+++ src/src/dof.cpp
@@ -3,7 +3,7 @@
 
 unsigned int Dof::global_dof_id_counter = 0;
 unsigned int Dof::get_global_dof_id_counter(){return global_dof_id_counter;};
-float Dof::get_value(){return value;};
+double Dof::get_value(){return value;};
 
 Dof::~Dof(){
     --global_dof_id_counter;
